Submitting my solutions
It was quite the atmosphere in the room at the Code100 final! When I
spotted that you'd released the challenges, I had to have a crack.

These were my first attempts that worked. I tried to write them in the
spirit of the competition. I couldn't imagine doing so on stage with an
enormous audience staring you down.